[fix]: [PIE-7095]: Delete nodeExecutions and planExecutions on piepline delete (#43330)
* [fix]: [PIE-7095]: Adding support for pipeline-metadata delete

* [fix]: [PIE-7095]: Handling pipelinemetadata delete

* [fix]: [PIE-7095]: Adding support for execution delete on pipeline delete

* [fix]: [PIE-7095]: Adding support for execution delete on pipeline delete

* [fix]: [PIE-7095]: Added unit test for Account and Pipeline Stream listeners

* [fix]: [PIE-7095]: Added unit test for BarrierServiceImplTest delete method

* [fix]: [PIE-7095]: Added unit test for PmsExecutionSummaryReadHelperTest iterator method

* [fix]: [PIE-7095]: Adding Unit test for PmsExecutionSummaryReadHelperTest

* [fix]: [PIE-7095]: Adding Unit test

* [fix]: [PIE-7095]: Adding AutoLogContext

* [fix]: [PIE-7095]: Added deletion of sweeping output

* [fix]: [PIE-7095]: Added deletion of outcomes

* [fix]: [PIE-7095]: Added deletion of interrupts

* [fix]: [PIE-7095]: Updated unit tests

* [fix]: [PIE-7095]: Updated unit tests

* [fix]: [PIE-7095]: Added deleted input set observer files

* [fix]: [PIE-7095]: Added deleted input set observer files

* [fix]: [PIE-7095]: Added unit test fix

* [fix]: [PIE-7095]: Added unit test fix

* [fix]: [PIE-7095]: Added Deletion of graph metadata

* [fix]: [PIE-7095]: Added Deletion of graph metadata

* [fix]: [PIE-7095]: Added analytics node for planExecution dashboard

* [fix]: [PIE-7095]: Remove unnecessary method in StepDetailInstance

* [fix]: [PIE-7095]: Added WaitInstance delete and its unit test for all correlationIds equal to nodeExecution

* [fix]: [PIE-7095]: Added TimeoutEngine delete and its unit test for all timeoutInstanceIds

* [fix]: [PIE-7095]: Added NodeExecutionService delete metadata

* [fix]: [PIE-7596]: Fix unit tests

* [fix]: [PIE-7596]: Add NotifyResponses delete

* [fix]: [PIE-7095]: Modify NodeExecution Delete metadata

* [fix]: [PIE-7095]: Add NodeExecutionDeleteObserver to delete its metadata in sync thread but in respective module and folder structure

* [fix]: [PIE-7095]: Fix tests

* [fix]: [PIE-7095]: Add handling of NodeEntity deletion

* [fix]: [PIE-7095]: Adding delete for nodeExecutionInfo on nodeExecutionDelete

* [fix]: [PIE-7095]: Adding delete for WaitStepService on nodeExecutionDelete

* [fix]: [PIE-7095]: Adding delete for ExecutionInputService on nodeExecutionDelete

* [fix]: [PIE-7095]: Adding delete for nodeExecutions

* [fix]: [PIE-7095]: Adding delete for PlanExecutionService

* [fix]: [PIE-7708]: Adding prefix env topic name in streams

* [fix]: [PIE-7095]: Adding unit tests

* [fix]: [PIE-7095]: Adding unit tests

* [fix]: [PIE-7095]: Adding unit tests and delete for planExecutions

* [fix]: [PIE-7095]: Adding unit tests

+++ pipeline-service/modules/orchestration-steps/src/main/java/io/harness/steps/resourcerestraint/service/ResourceRestraintServiceImpl.java
@@ -8,9 +8,11 @@
 package io.harness.steps.resourcerestraint.service;
 
 import static io.harness.exception.WingsException.USER;
+import static io.harness.springdata.PersistenceUtils.DEFAULT_RETRY_POLICY;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.data.structure.EmptyPredicate;
 import io.harness.exception.InvalidRequestException;
 import io.harness.repositories.ResourceRestraintRepository;
 import io.harness.steps.resourcerestraint.beans.ResourceRestraint;
@@ -19,6 +21,7 @@ import com.google.inject.Inject;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
+import net.jodah.failsafe.Failsafe;
 import org.springframework.dao.DuplicateKeyException;
 
 @OwnedBy(HarnessTeam.PIPELINE)
@@ -53,6 +56,12 @@ public class ResourceRestraintServiceImpl implements ResourceRestraintService {
 
   @Override
   public void deleteAllRestraintForGivenAccount(String accountId) {
-    resourceRestraintRepository.deleteByAccountId(accountId);
+    if (EmptyPredicate.isEmpty(accountId)) {
+      return;
+    }
+    Failsafe.with(DEFAULT_RETRY_POLICY).get(() -> {
+      resourceRestraintRepository.deleteByAccountId(accountId);
+      return true;
+    });
   }
 }
